separated geojson parse logic to separate class
